reset frontend yarn changes

+++ studio/app/optinist/wrappers/caiman/motion_correction.py
@@ -95,7 +95,7 @@ def __process_images(images):
         .transpose(2, 0, 1)
     )
 
-    rois = rois.astype(np.float)
+    rois = rois.astype(np.float32)
     for i, _ in enumerate(rois):
         rois[i] *= i + 1
